Fix: modify directory
